- Removed AlloySelectionGrid and integrated it directly into the page, since they are tightly coupled in terms of functionality, and the component provides no reusability.
